+++ CHANGELOG.md
@@ -5,6 +5,16 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+Added:
+
+  * METS Server: providing concurrent additional access to the METS file for parallel processing, #966
+  * Web API: Cache jobs in the processing server with an optional callback once processed, #1069
+  * Web API: Lock pages output file groups of a workspace to prevent simultaneous non-additive access to workspaces, #1069
+  * Web API: Support job dependency for caching complete fully-deterministic workflows, #1069
+  * Web API: Processing server will start all ready requests, not just the first one, #1069
+
+## [2.53.0] - 2023-08-21
+
 Fixed:
 
   * `WorkspaceValidator`: make the check for consistency of `pc:Page[@pcGtsId]` and `mets:file[@ID]` optional with the `mets_fileid_page_pcgtsid` skip flag, #1066
@@ -1793,6 +1803,7 @@ Fixed
 Initial Release
 
 <!-- link-labels -->
+[2.53.0]: ../../compare/v2.53.0..v2.52.0
 [2.52.0]: ../../compare/v2.52.0..v2.51.0
 [2.51.0]: ../../compare/v2.51.0..v2.50.0
 [2.50.0]: ../../compare/v2.50.0..v2.49.0